+++ dbt/adapters/databricks/connections.py
@@ -1,3 +1,4 @@
+import json
 import warnings
 from contextlib import contextmanager
 from dataclasses import dataclass
@@ -53,6 +54,7 @@ DBR_VERSION_REGEX = re.compile(r"([1-9][0-9]*)\.(x|0|[1-9][0-9]*)")
 DBT_DATABRICKS_INVOCATION_ENV = "DBT_DATABRICKS_INVOCATION_ENV"
 DBT_DATABRICKS_INVOCATION_ENV_REGEX = re.compile("^[A-z0-9\\-]+$")
 EXTRACT_CLUSTER_ID_FROM_HTTP_PATH_REGEX = re.compile(r"/?sql/protocolv1/o/\d+/(.*)")
+DBT_DATABRICKS_HTTP_SESSION_HEADERS = "DBT_DATABRICKS_HTTP_SESSION_HEADERS"
 
 
 @dataclass
@@ -450,6 +452,33 @@ class DatabricksConnectionManager(SparkConnectionManager):
             )
 
     @classmethod
+    def get_all_http_headers(
+        cls, user_http_session_headers: Dict[str, str]
+    ) -> List[Tuple[str, str]]:
+        http_session_headers_str: Optional[str] = os.environ.get(
+            DBT_DATABRICKS_HTTP_SESSION_HEADERS
+        )
+
+        http_session_headers_dict: Dict[str, str] = (
+            {k: json.dumps(v) for k, v in json.loads(http_session_headers_str).items()}
+            if http_session_headers_str is not None
+            else {}
+        )
+
+        intersect_http_header_keys = (
+            user_http_session_headers.keys() & http_session_headers_dict.keys()
+        )
+
+        if len(intersect_http_header_keys) > 0:
+            raise dbt.exceptions.ValidationException(
+                f"Intersection with reserved http_headers in keys: {intersect_http_header_keys}"
+            )
+
+        http_session_headers_dict.update(user_http_session_headers)
+
+        return list(http_session_headers_dict.items())
+
+    @classmethod
     def open(cls, connection: Connection) -> Connection:
         if connection.state == ConnectionState.OPEN:
             logger.debug("Connection is already open, skipping open.")
@@ -474,8 +503,9 @@ class DatabricksConnectionManager(SparkConnectionManager):
         cls.validate_creds(creds, required_fields)
 
         connection_parameters = creds.connection_parameters.copy()  # type: ignore[union-attr]
-        http_headers: List[Tuple[str, str]] = list(
-            connection_parameters.pop("http_headers", {}).items()
+
+        http_headers: List[Tuple[str, str]] = cls.get_all_http_headers(
+            connection_parameters.pop("http_headers", {})
         )
 
         for i in range(1 + creds.connect_retries):